+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # 3.5.1
 - dezog.run: Fixed extensions. Now also .p81 and .81 files are supported (ZX81).
 - zsim restore: In case of an error during "restore" a rollback is done.
+- Fixed #134: ZX81 keyboard: highlighting of keys is shifted
 
 # 3.5.0
 - zsim: