+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+# 3.5.2
+- If there is a problem with a logpoint it is now shown as warning to the user. Before it was silently ignored.
+
 # 3.5.1
 - dezog.run: Fixed extensions. Now also .p81 and .81 files are supported (ZX81).
 - zsim restore: In case of an error during "restore" a rollback is done.